feat(ui): edit companion avatar from the edit form, not the dashboard
Move the editable avatar (file upload + Immich pick) to the profile card
in the edit form. The dashboard hero avatar is now display-only; tapping
still opens the lightbox when a photo exists.
