+++ src/routes/(app)/(companion)/[companionId]/+page.svelte
@@ -2,9 +2,6 @@
 	import type { PageData } from './$types';
 	import CompanionAvatar from '$lib/components/CompanionAvatar.svelte';
 	import WeightSparkline from '$lib/components/WeightSparkline.svelte';
-	import ImmichPicker from '$lib/components/ImmichPicker.svelte';
-	import { bustAvatarCache } from '$lib/avatarCache.svelte';
-	import { invalidateAll } from '$app/navigation';
 	import LocalTime from '$lib/components/LocalTime.svelte';
 	import ByLine from '$lib/components/ByLine.svelte';
 	import { localDateISO } from '$lib/date';
@@ -137,41 +134,6 @@
 		avatarLightboxOpen = false;
 		(document.activeElement as HTMLElement)?.blur();
 	}
-
-	// Immich avatar picker
-	let immichAvatarPickerOpen = $state(false);
-	let immichAvatarError = $state('');
-	let immichAvatarErrorTimer: ReturnType<typeof setTimeout>;
-
-	function setImmichAvatarError(msg: string) {
-		immichAvatarError = msg;
-		clearTimeout(immichAvatarErrorTimer);
-		immichAvatarErrorTimer = setTimeout(() => (immichAvatarError = ''), 5000);
-	}
-
-	async function pickAvatarFromImmich(assetId: string) {
-		try {
-			const res = await fetch(`/api/companions/${companion.id}/avatar/from-immich`, {
-				method: 'POST',
-				headers: { 'Content-Type': 'application/json' },
-				body: JSON.stringify({ assetId })
-			});
-			if (!res.ok) {
-				const err = await res
-					.json()
-					.catch(() => ({ message: t(locale, 'immich.picker.pickFailed') }));
-				setImmichAvatarError(err.message ?? t(locale, 'immich.picker.pickFailed'));
-				return;
-			}
-			bustAvatarCache(companion.id);
-			immichAvatarPickerOpen = false;
-			await invalidateAll();
-		} catch {
-			setImmichAvatarError(t(locale, 'immich.picker.pickFailed'));
-		}
-	}
-
-	$effect(() => () => clearTimeout(immichAvatarErrorTimer));
 
 	// Pending reminder dismissals
 	const undoDelayMs = $derived((data.reminderUndoSeconds ?? 0) * 1000);
@@ -576,10 +538,7 @@
 					avatarPath={companion.avatarPath}
 					name={companion.name}
 					size="lg"
-					editable={companion.isActive}
 					onlightbox={avatarUrl ? () => (avatarLightboxOpen = true) : undefined}
-					immichEnabled={data.immichEnabled}
-					onpickImmich={() => (immichAvatarPickerOpen = true)}
 				/>
 				<div class="flex-1 min-w-0">
 					<div class="flex items-start justify-between gap-2">
@@ -956,20 +915,3 @@
 		</Card>
 	{/if}
 </div>
-
-{#if data.immichEnabled}
-	<ImmichPicker
-		open={immichAvatarPickerOpen}
-		onpick={pickAvatarFromImmich}
-		onclose={() => (immichAvatarPickerOpen = false)}
-	/>
-{/if}
-
-{#if immichAvatarError}
-	<div
-		role="alert"
-		class="fixed bottom-4 right-4 z-50 max-w-sm rounded-lg bg-destructive text-destructive-foreground px-4 py-2 text-sm shadow-lg"
-	>
-		{immichAvatarError}
-	</div>
-{/if}
